Move predicate search to new ListUtil class
